Remove special handling of ThisExpressions
This logic was added during a time where we still wanted to make reflective-bind work for all callbacks, even ones that render. However, we have since given in to the fact that there are too many edge cases to consider for render callbacks that can cause stale renders, so reflective-bind should not be run on `render` props, which means that any special optimization of ThisExpressions is pointless.